requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
streamlit==1.37.0
scikit-learn==1.5.1
google-generativeai==0.5.4
tqdm==4.66.4
openai==1.3.7 
httpx==0.24.1

